anpr-backend/requirements.txt
# Flask Core
Flask==3.0.0
Werkzeug==3.0.1

# API Extensions
Flask-RESTful==0.3.10
Flask-CORS==4.0.0

# Configuration & Environment
python-dotenv==1.0.0

# Image Processing (Core dependencies - ML libraries added later)
Pillow==10.4.0
#numpy==2.0.0

# HTTP Client
requests==2.31.0

# Logging
python-json-logger==2.0.7

# Validation
marshmallow==3.20.1

# Utilities
python-dateutil==2.8.2

# Production Server
gunicorn==21.2.0

# Development (optional)
# pytest==7.4.3
# pytest-flask==1.3.0
# black==23.11.0
# flake8==6.1.0

# ML Dependencies (to be added when implementing detection)
# opencv-python==4.8.1.78
# ultralytics==8.0.196  # YOLOv8
# easyocr==1.7.0
# pytesseract==0.3.10
